backend/requirements.txt
asgiref==3.5.0
Django==3.2.19
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.14
Pillow==9.2.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-decouple==3.8
whitenoise==6.5
gunicorn==20.1.0